More signal managment. More or less correct behavior for ^C and ^D

+++ src/minishell.c
@@ -6,7 +6,7 @@
 /*   By: ejafer <ejafer@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/28 20:35:09 by ejafer            #+#    #+#             */
-/*   Updated: 2022/06/21 02:40:04 by ejafer           ###   ########.fr       */
+/*   Updated: 2022/06/21 15:30:37 by ejafer           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,6 +17,8 @@ void	minishell(t_mini *mini)
 	while (1)
 	{
 		mini->line = readline("\033[32mMinishell\033[0m$: ");
+		if (!mini->line)
+			break ;
 		add_history(mini->line);
 		split_line(mini);
 		parse(mini);
@@ -24,27 +26,52 @@ void	minishell(t_mini *mini)
 	}
 }
 
-void	sighandler_exit(int n)
+void	sighandler_stop(int signo)
 {
-	signal(SIGQUIT, SIG_IGN);
-	kill(0, SIGQUIT);
-	signal(SIGQUIT, SIG_DFL);
-	while (waitpid(-1, NULL, 0) > 0)
+	if (signo == SIGTSTP)
 	{
+		exit(128 + SIGTSTP);
 	}
-	exit(0);
+}
+
+void	sighandler_interupt(int signo)
+{
+	if (signo == SIGINT)
+	{
+		rl_point = 0;
+		rl_mark = 0;
+		rl_end = 0;
+		rl_line_buffer[0] = 0;
+		printf("\n");
+		rl_on_new_line();
+		rl_redisplay();
+	}
+}
+
+void	sighandler_exit(int signo)
+{
+	if (signo == SIGUSR2)
+	{
+		signal(SIGQUIT, SIG_IGN);
+		kill(0, SIGQUIT);
+		signal(SIGQUIT, SIG_DFL);
+		while (waitpid(-1, NULL, 0) > 0)
+		{
+		}
+		exit(0);
+	}
 }
 
 int	main(int argc, char **argv, char **env)
 {
 	t_mini	*mini;
 
-	signal(SIGINT, SIG_IGN);
+	rl_catch_signals = 1;
+	signal(SIGINT, sighandler_interupt);
+	signal(SIGTSTP, sighandler_stop);
 	signal(SIGUSR2, sighandler_exit);
-	//signal(SIGTSTP, signal_handler);
 	mini = malloc(sizeof(t_mini));
 	mini->pid = getpid();
-	printf("pid: %d\n", mini->pid);
 	mini->heredocid = 0;
 	mini->env = env;
 	mini->argc = argc;
